+++ src/Config/IslandoraSolrMetadataFieldConfig.php
@@ -4,6 +4,8 @@ namespace Drupal\islandora_solr_metadata\Config;
 
 use Drupal\Core\Config\ConfigFactoryInterface;
 use Drupal\Component\Utility\SortArray;
+use Drupal\Core\Cache\RefineableCacheableDependencyInterface;
+use Drupal\Core\Cache\RefineableCacheableDependencyTrait;
 
 /**
  * Metadata display profile field config.
@@ -11,7 +13,14 @@ use Drupal\Component\Utility\SortArray;
  * Provides implementing code a method of accessing field info by using
  * actual Solr field names.
  */
-class IslandoraSolrMetadataFieldConfig {
+class FieldConfig implements FieldConfigInterface, RefineableCacheableDependencyInterface {
+  use FieldConfigTrait;
+  use RefineableCacheableDependencyTrait;
+
+  /**
+   * Defines the replacement string for periods in field names.
+   */
+  const PERIOD_REPLACEMENT = '~dot~';
 
   /**
    * Islandora Solr Metadata configs object.
@@ -28,19 +37,11 @@ class IslandoraSolrMetadataFieldConfig {
    */
   public function __construct(ConfigFactoryInterface $config_factory) {
     $this->config = $config_factory->getEditable('islandora_solr_metadata.configs');
+    $this->addCacheableDependency($this->config);
   }
 
   /**
-   * Gets the fields property of this configuration.
-   *
-   * Machine names of keys are replaced with the actual field names.
-   *
-   * @param string $config_name
-   *   The name of the configuration to get fields for.
-   *
-   * @return array
-   *   An associative array mapping Solr field names to their configurations.
-   *   These will come back sorted by weight.
+   * {@inheritdoc}
    */
   public function getFields($config_name) {
     $fields = [];
@@ -54,15 +55,7 @@ class IslandoraSolrMetadataFieldConfig {
   }
 
   /**
-   * Gets an individual field config from this configuration.
-   *
-   * @param string $field_name
-   *   The name of the field to get.
-   * @param string $config_name
-   *   The name of the configuration to get the field config from.
-   *
-   * @return array|null
-   *   The field config, or NULL if no such field exists.
+   * {@inheritdoc}
    */
   public function getField($field_name, $config_name) {
     $machine_name = static::getMachineName($field_name);
@@ -70,12 +63,7 @@ class IslandoraSolrMetadataFieldConfig {
   }
 
   /**
-   * Adds or overwrites the fields for a config with a provided set.
-   *
-   * @param array $fields
-   *   Formatted field info to set in this config.
-   * @param string $config_name
-   *   The configuration to set.
+   * {@inheritdoc}
    */
   public function setFields(array $fields, $config_name) {
     foreach ($fields as $field => $config) {
@@ -84,14 +72,7 @@ class IslandoraSolrMetadataFieldConfig {
   }
 
   /**
-   * Adds or overwrites a field configuration.
-   *
-   * @param string $field_name
-   *   The name of the Solr field to set the config for.
-   * @param array $config
-   *   The configuration to use for this field.
-   * @param string $config_name
-   *   The configuration this field is being set for.
+   * {@inheritdoc}
    */
   public function setField($field_name, array $config, $config_name) {
     $machine_name = static::getMachineName($field_name);
@@ -111,17 +92,14 @@ class IslandoraSolrMetadataFieldConfig {
         'min_wordsafe_length' => isset($config['truncation']['ellipsis']) ? $config['truncation']['ellipsis'] : 0,
       ],
     ];
-    $this->config->set("configs.$config_name.fields.$machine_name", $new_config)
+
+    $this->config
+      ->set("configs.$config_name.fields.$machine_name", $new_config)
       ->save();
   }
 
   /**
-   * Removes the given fields from the config, if they exist.
-   *
-   * @param array $fields
-   *   An array of fields to remove.
-   * @param string $config_name
-   *   The configuration to set.
+   * {@inheritdoc}
    */
   public function deleteFields(array $fields, $config_name) {
     foreach ($fields as $field) {
@@ -132,66 +110,13 @@ class IslandoraSolrMetadataFieldConfig {
   }
 
   /**
-   * Gets an empty config using the default settings.
-   *
-   * @return array
-   *   An empty Solr metadata display config.
-   */
-  public static function getEmptyConfig() {
-    return [
-      'cmodel_associations' => [],
-      'description' => [
-        'description_field' => '',
-        'description_label' => '',
-        'truncation' => static::getEmptyTruncation(),
-      ],
-      'fields' => [],
-    ];
-  }
-
-  /**
-   * Gets an empty field using the default settings.
-   *
-   * @return array
-   *   An empty Solr metadata display field config.
-   */
-  public static function getEmptyField() {
-    return [
-      'weight' => 0,
-      'display_label' => '',
-      'hyperlink' => FALSE,
-      'uri_replacement' => FALSE,
-      'date_format' => '',
-      'enable_permissions' => FALSE,
-      'permissions' => [],
-      'truncation' => static::getEmptyTruncation(),
-    ];
-  }
-
-  /**
-   * Gets an empty truncation configuration using the default settings.
-   *
-   * @return array
-   *   An empty truncation configuration.
-   */
-  public static function getEmptyTruncation() {
-    return [
-      'truncation_type' => 'separate_value_option',
-      'max_length' => 0,
-      'word_safe' => FALSE,
-      'ellipsis' => FALSE,
-      'min_wordsafe_length' => 0,
-    ];
-  }
-
-  /**
    * Defines the replacement string for periods in field names.
    *
    * @return string
    *   The string to use to replace periods in machine names. This obviously
    *   should not contain periods; if it does, your configs may blow up.
    */
-  public static function getPeriodReplacement() {
+  protected static function getPeriodReplacement() {
     return '~dot~';
   }
 
@@ -204,8 +129,8 @@ class IslandoraSolrMetadataFieldConfig {
    * @return string
    *   The machine name for this field.
    */
-  public static function getMachineName($field_name) {
-    return str_replace('.', static::getPeriodReplacement(), $field_name);
+  protected static function getMachineName($field_name) {
+    return str_replace('.', static::PERIOD_REPLACEMENT, $field_name);
   }
 
   /**
@@ -218,7 +143,7 @@ class IslandoraSolrMetadataFieldConfig {
    *   The name of the Solr field.
    */
   protected function getFieldName($machine_name) {
-    return str_replace(static::getPeriodReplacement(), '.', $machine_name);
+    return str_replace(static::PERIOD_REPLACEMENT, '.', $machine_name);
   }
 
 }
